[refactor]:フォルダ構造の整理
