+++ src/components/Join/MainVisual.js
@@ -1,5 +1,7 @@
 import React from "react";
-import ReadImage from './ReadImage';
+
+// components
+import ReadImage from '../ReadImage';
 
 export default function JoinMainVisual() {
     const image_path = "./image/2.jpg";
