+++ src/components/News/index.js
@@ -1,8 +1,12 @@
 import React from 'react';
-import './font.css'
-import Header from "./Header";
-import Fotter from "./Footer";
-import Right from "./Right";
+
+// style
+import '../../style/font.css'
+
+// components
+import Header from "../Header";
+import Fotter from "../Footer";
+import Right from "../Right";
 
 export default function Activities(){
     return (
