+++ src/components/Right/index.js
@@ -2,7 +2,7 @@ export default function Right(){
     return (
         <>
         <section className="text-center text-[10px]  md:bg-[#999] bg-[#fff]">
-        <div className='md:text-[#fff] text-[#000]'>© 2024 コンテンツクリエーションサークル. All Rights Reserved.</div>
+        <div className='md:text-[#fff]'>© 2024 コンテンツクリエーションサークル. All Rights Reserved.</div>
         </section>
         </>
     );
