--- src/font.css
@@ -1,29 +1,0 @@
-@font-face {
-    font-family: Museno-100;
-    src: url(../public/font/Museno-Sans-Cond/100.otf);
-}
-@font-face {
-    font-family: Museno-300;
-    src: url(../public/font/Museno-Sans-Cond/300.otf);
-}
-@font-face {
-    font-family: Museno-500;
-    src: url(../public/font/Museno-Sans-Cond/500.otf);
-}
-@font-face {
-    font-family: Museno-700;
-    src: url(../public/font/Museno-Sans-Cond/700.otf);
-}
-@font-face {
-    font-family: Museno-900;
-    src: url(../public/font/Museno-Sans-Cond/900.otf);
-}
-@font-face {
-    font-family: Noto-Sans-Bold;
-    src: url(../public/font/Noto-Sans-JP/NotoSansJP-Bold.ttf);
-}
-@font-face {
-    font-family: Noto-Sans-Regular;
-    src: url(../public/font/Noto-Sans-JP/NotoSansJP-Regular.ttf);
-}
-
