+++ src/style/headerAnim.css
@@ -1,8 +1,6 @@
-
-/*========= ボタンのためのCSS ===============*/
 .openbtn{
 	position:fixed;
-    z-index: 9999;/*ボタンを最前面に*/
+    z-index: 9999;
 	top:10px;
 	right: 10px;
 	cursor: pointer;
@@ -10,7 +8,6 @@
     height:50px;
 }
 	
-/*×に変化*/	
 .openbtn span{
     display: inline-block;
     transition: all .4s;
